requirements.txt
pymysql>=1.1.1
sqlalchemy>=2.0.36
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn>=0.30.0
git+https://github.com/runekaagaard/mcp-alchemy.git